Convert models to postgres
