+++ database/models/userModel.js
@@ -1,4 +1,4 @@
-const db = require('../../data/dbConfig');
+const db = require("../dbConfig");
 
 module.exports = {
   get,
@@ -10,37 +10,33 @@ module.exports = {
 };
 
 function get() {
-  return db('users');
+  return db("users");
 }
 
 function getById(id) {
-  return db('users')
+  return db("users")
     .where({ id })
     .first();
 }
 
 function getByUsername(username) {
-  return db('users')
+  return db("users")
     .where({ username })
     .first();
 }
 
 function insert(user) {
-  return db('users')
-    .insert(user)
-    .then(ids => {
-      return getById(ids[0]);
-    });
+  return db("users").insert(user, "id");
 }
 
 function update(id, changes) {
-  return db('users')
+  return db("users")
     .where({ id })
     .update(changes);
 }
 
 function remove(id) {
-  return db('users')
-    .where('id', id)
+  return db("users")
+    .where({ id })
     .del();
 }
